Add components folder and move App.tsx into src folder
