+++ src/App.tsx
@@ -1,15 +1,13 @@
 import React from 'react';
 import { createStackNavigator } from '@react-navigation/stack';
 import { NavigationContainer } from '@react-navigation/native';
-import Header from './src/screens/shared/header'
-import Home from './src/screens/home';
+
+import Header from './components/header'
+import Home from './screens/home';
 
 const App = () => {
-
   const Stack = createStackNavigator();
 
-
-  
   return (
     <NavigationContainer>
       <Header />
